+++ tutorials/tutorial3.py
@@ -1,24 +1,23 @@
-""" This example shows how to simulate heat diffusion in a hollow sphere made out of ATS58 (PCM material). """
+""" This example shows how to simulate heat diffusion in a sphere made out of swedish diabase stone. """
 
 import openterrace
 
-Ri = 0.005
-Ro = 0.025
+R = 0.025
 T_init = 40+273.15
 T_room = 80+273.15
-h = 50
+h = 200
 
-ot = openterrace.Simulate(t_end=7200, dt=0.05, sim_name='tutorial3')
+ot = openterrace.Simulate(t_end=15*60, dt=0.01, sim_name='tutorial3')
 
-bed = ot.createPhase(n=30, type='bed')
-bed.select_substance(substance='ATS58')
-bed.select_domain_shape(domain='hollow_sphere_1d', Rinner=Ri, Router=Ro)
+bed = ot.createPhase(n=50, type='bed')
+bed.select_substance('swedish_diabase')
+bed.select_domain_shape(domain='sphere_1d', R=R)
 bed.select_schemes(diff='central_difference_1d')
 bed.select_initial_conditions(T=T_init)
 bed.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), 0))
 bed.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), -1))
-bed.select_source_term(source_type='thermal_resistance', R=1/(h*4*3.14159*Ro**2), T_inf=T_room, position=(slice(None, None, None), -1))
-bed.select_output(times=range(0, 7200, 300), parameters=['T'])
+bed.select_source_term(source_type='thermal_resistance', R=1/(h*4*3.14159*R**2), T_inf=T_room, position=(slice(None, None, None), -1))
+bed.select_output(times=range(0, 15*60+60, 60), parameters=['T'])
 
 ot.run_simulation()
 ot.generate_plot(pos_phase=bed, data_phase=bed)
