### 2023-12-11 #6128 @danielskatz
@jakobhaervig - thanks for your submission.

1. Note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

2. I think this should be in another track, if it goes forward. It's now in 1) Computer science, Information Science, and Mathematics, and the two options that seem most appropriate to me are 2) Biomedical Engineering, Biosciences, Chemistry, and Materials or 3) Physics and Engineering.  Can you let me know which of these you think best fits the work?

3. Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.
