### 2023-12-15 3ae62a9
Added "_update_massflow" to the phase class
