### 2024-01-02 7ca134e
Update of plotting
