+++ tutorials/tutorial5.py
@@ -14,7 +14,8 @@ fluid.select_substance(substance='air')
 fluid.select_domain_shape(domain='cylinder_1d', D=0.3, H=1)
 fluid.select_porosity(phi=0.4)
 fluid.select_schemes(diff='central_difference_1d', conv='upwind_1d')
-fluid.select_initial_conditions(T=273.15+25, mdot=0.001)
+fluid.select_initial_conditions(T=273.15+25)
+fluid.select_massflow(mdot=0.001)
 fluid.select_bc(bc_type='fixedValue',
                    parameter='T',
                    position=(slice(None, None, None), 0),
