+++ tutorials/tutorial7.py
@@ -9,7 +9,8 @@ fluid.select_substance(substance='water')
 fluid.select_domain_shape(domain='cylinder_1d', D=0.1, H=1)
 fluid.select_porosity(phi=0.4)
 fluid.select_schemes(diff='central_difference_1d', conv='upwind_1d')
-fluid.select_initial_conditions(T=273.15+20, mdot=0.01)
+fluid.select_initial_conditions(T=273.15+20)
+fluid.select_massflow(mdot=0.01)
 fluid.select_bc(bc_type='fixedValue', parameter='T', position=(slice(None, None, None), 0), value=273.15+80)
 fluid.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), -1))
 fluid.select_output(times=[30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 360, 390, 420, 450, 480, 510, 540, 570, 600, 900, 1800, 3600, 5400], parameters=['T'])
