+++ tutorials/tutorial3.py
@@ -17,8 +17,8 @@ bed.select_initial_conditions(T=T_init)
 bed.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), 0))
 bed.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), -1))
 bed.select_source_term(source_type='thermal_resistance', R=1/(h*4*3.14159*R**2), T_inf=T_room, position=(slice(None, None, None), -1))
-bed.select_output(times=range(0, 15*60+60, 60), parameters=['T'])
+bed.select_output(times=range(0, 15*60+60, 60))
 
 ot.run_simulation()
-ot.generate_plot(pos_phase=bed, data_phase=bed)
-ot.generate_animation(pos_phase=bed, data_phase=bed)
+ot.generate_plot(x=bed.node_pos, y=bed.data.T, times=bed.data.time, xlabel='Position', ylabel='Temperature')
+ot.generate_animation(x=bed.node_pos, y=bed.data.T, times=bed.data.time, xlabel='Position', ylabel='Temperature')
