+++ openterrace/openterrace.py
@@ -77,27 +77,10 @@ class Simulate:
             if self.flag_coupling:
                 self._coupling()
 
-    def generate_plot(self, x:object=None, y:object=None, parameter:str='T'):
-        filename='ot_plot_'+self.sim_name+'_'+str(x)+'_'+str(y)+'_'+parameter+'.png'
-
-        # if pos_phase == data_phase:
-        #     y = np.mean(getattr(data_phase.data, parameter),1)
-        # else:
-        #     y = np.mean(getattr(data_phase.data, parameter),2)
-        
-        # if y.shape[1] == 1:
-        #     y = np.append(y, y, 1)
-
-        if len(x.shape) == 1:
-            xp = x
-
-        print(len(x.shape),x.shape)
-        print(len(y.shape),y.shape)
-
-        sys.exit()
-
+    def generate_plot(self, x:list[float]=None, y:list[float]=None, times:list[float]=None, xlabel:str=None, ylabel:str=None, name:str=None):
+        filename='ot_plot_'+self.sim_name+'_'+name+'.png'
         fig, axes = plt.subplots()
-        for i,time in enumerate(getattr(data_phase.data, 'time')):
+        for i,time in enumerate(times):
             timelabel = u'$%s$' % time
             plt.plot(x, y[i,:].transpose()-273.15, label=timelabel)
 
@@ -105,37 +88,27 @@ class Simulate:
         labelLines(lines, fontsize=8, align=True)
 
         plt.grid()
-        plt.xlabel('Position (m)')
-        plt.ylabel(u'$%s_{%s}$ (\u00B0C)' % (parameter, data_phase.type))
+        plt.xlabel('%s' % (xlabel))
+        plt.ylabel('%s' % (ylabel))
         plt.savefig(filename)
 
-    # def generate_animation(self, pos_phase:object=None, data_phase:object=None, parameter:str='T'):
-    #     def _update(frame, parameter):
-    #         x = pos_phase.domain.node_pos
-    #         if pos_phase == data_phase:
-    #             y = np.mean(getattr(data_phase.data, parameter),1)
-    #         else:
-    #             y = np.mean(getattr(data_phase.data, parameter),2)
-    #         times = getattr(data_phase.data, 'time')
+    def generate_animation(self, x:list[float]=None, y:list[float]=None, times:list[float]=None, xlabel:str=None, ylabel:str=None):
+        def _update(frame, y, times):       
+            ax.clear()
+            ax.set_xlabel(xlabel)
+            ax.set_ylabel(ylabel)
+            ax.set_xlim(np.min(x), np.max(x))
+            ax.set_ylim(np.min(y)-0.05*np.max(y), np.max(y)+0.05*np.max(y))
+            ax.grid()
+            ax.plot(x, y[frame,:].transpose(), color = '#4cae4f')
+            ax.set_title('Time: ' + str(np.round(times[frame], decimals=2)) + ' s')
 
-    #         if y.shape[1] == 1:
-    #             y = np.append(y, y, 1)
-                
-    #         ax.clear()
-    #         ax.set_xlabel('Position (m)')
-    #         ax.set_ylabel(u'$%s_{%s}$ (\u00B0C)' % (parameter, data_phase.type))
-    #         ax.set_xlim(np.min(pos_phase.domain.node_pos), np.max(pos_phase.domain.node_pos))
-    #         ax.set_ylim(np.min(getattr(data_phase.data,parameter)-273.15)-0.05*(np.max(getattr(data_phase.data, parameter)-273.15)), np.max(getattr(data_phase.data, parameter)-273.15)+0.05*(np.max(getattr(data_phase.data, parameter)-273.15)))
[…]

[…]
