+++ tutorials/validation_grabo.py
@@ -9,7 +9,6 @@ ot = openterrace.Simulate(t_end=600*60, dt=0.1, sim_name='validation_grabo')
 fluid = ot.createPhase(n=290, type='fluid')
 fluid.select_substance_on_the_fly(rho=1000, cp=4200, k=0.6)
 fluid.select_domain_shape(domain='block_1d', A=0.734**2, L=1.45)
-#fluid.select_domain_shape(domain='cylinder_1d', D=0.79, H=1.9)
 fluid.select_porosity(phi=0.65)
 fluid.select_schemes(diff='central_difference_1d', conv='upwind_1d')
 fluid.select_initial_conditions(T=273.15+47)
