### 2024-02-07 #6128 @AdamRJensen
@jakobhaervig could you search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and and see if any are suitable for reviewing this submission? If so, please mention them here in this thread (without tagging them with an @).

@danielskatz of course... I'm a new editor here, my apologies 😄
