### 2024-02-07 #6128 @AdamRJensen
:wave: @hewei8622, @mdoucet, I see that you both are signed up as JOSS reviewers with the tag "energy storage" and was wondering if any of you be available to review this submission?
