### 2024-02-14 2d3adfd
Add V0 function to calculate volume of shape
