### 2024-02-14 af9a489
Refactor simulation code and update plot generation
