+++ tutorials/tutorial2.py
@@ -1,8 +1,12 @@
-""" This example shows how to simulate advection of temperature in a cylindrical tank without any bed material. """
+""" 
+This example shows how to simulate advection of temperature
+in a cylindrical tank without any bed material.
+"""
 
 import openterrace
+import matplotlib.pyplot as plt
 
-ot = openterrace.Simulate(t_end=600, dt=0.01, sim_name='tutorial2')
+ot = openterrace.Simulate(t_end=600, dt=0.01)
 
 fluid = ot.createPhase(n=500, type='fluid')
 fluid.select_substance_on_the_fly(rho=1000, cp=4200, k=0.6)
@@ -12,12 +16,28 @@ fluid.select_initial_conditions(T=273.15+100)
 fluid.select_massflow(mdot=[
                             [0, 0.1],
                             [400, 0.1],
-                            [600, -0.1]
+                            [500, -0.1]
                             ])
 
-fluid.select_bc(bc_type='fixedValue', parameter='T', position=(slice(None, None, None), 0), value=273.15+600)
-fluid.select_bc(bc_type='zeroGradient', parameter='T', position=(slice(None, None, None), -1), value=0)
-fluid.select_output(times=range(0, 10*60+60, 60))
+fluid.select_bc(bc_type='fixedValue', 
+                parameter='T',
+                position=(slice(None, None, None), 0),
+                value=273.15+600)
+fluid.select_bc(bc_type='zeroGradient', 
+                parameter='T',
+                position=(slice(None, None, None), -1),
+                value=0)
+fluid.select_output(times=range(0, 10*60+45, 45))
 
 ot.run_simulation()
-ot.generate_plot(x=fluid.node_pos, y=fluid.data.T, times=fluid.data.time, xlabel='Position', ylabel='Temperature')
+
+plt.plot(fluid.node_pos,fluid.data.T[:,0,:].T-273.15, label=fluid.data.time)
+plt.legend(title='Simulation time (s)')
+plt.show()
+plt.xlabel(u'Cylinder position (m)')
+plt.ylabel(u'Temperature (℃)')
+plt.grid()
+plt.grid(which='major', color='#DDDDDD', linewidth=1)
+plt.grid(which='minor', color='#EEEEEE', linestyle=':', linewidth=0.8)
+plt.minorticks_on()
+plt.savefig('ot_plot_tutorial2.png')

[…]
