### 2024-02-14 5e565ed
Refactor code to improve performance and readability
