+++ docs/user-guide/domains/lumped.md
@@ -1,0 +1,4 @@
+::: domains.lumped
+    options:
+      show_source: true
+      heading_level: 1
