### 2024-02-15 3542898
Update simulation start time and coupling heat transfer coefficient
