### 2024-02-21 cf63d19
Changed to constant properties of water
