### 2024-02-26 #6411 @editorialbot
[REVIEW]: OpenTerrace: A fast, flexible and extendable Python framework for thermal storage packed bed simulations
**Submitting author:** @jakobhaervig (<a href="http://orcid.org/0000-0001-8710-1610">Jakob Hærvig</a>)
**Repository:** https://github.com/OpenTerrace/openterrace-python
**Branch with paper.md** (empty if default branch): development
**Version:** v0.0.8
