+++ .github/workflows/tests.yml
@@ -9,17 +9,17 @@ jobs:
   test-pip:
     name: Run tests (with pip)
 
-    runs-on: ${{ matrix.os }}
-    
     strategy:
       matrix:
-        os: [ubuntu-latest]
-        python-version: ['3.11']
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: [3.8, 3.9, "3.10", "3.11", "3.12"]
 
     defaults:
       run:
         shell: bash
 
+    runs-on: ${{ matrix.os }}
+    
     steps:
     - uses: actions/checkout@v2
 
