### 2024-02-26 1b3a8d9
Update OpenTerrace documentation link
