### 2024-02-26 a79631a
Remove unused dependencies and profiling code
