+++ .github/workflows/docs.yml
@@ -1,4 +1,4 @@
-name: Docs Publish
+name: Generate Docs
 
 on:
   push:
