### 2024-02-26 5f1b9b9
Moved JOSS paper to its own folder
