### 2024-02-26 #6128 @ronshnapp
Hello Adam,

Apologies but I am totally booked these days, so I'll have to pass on this
one.

Best,
Ron

On Sat, Feb 24, 2024 at 1:39 AM Adam R. Jensen ***@***.***>
wrote:

> 👋 @ronshnapp <https://github.com/ronshnapp> and @dankusanovic
> <https://github.com/dankusanovic>, I see that you are signed up as JOSS
> reviewers and was wondering if any of you be available to review this
> submission?
>
