### 2024-02-26 a13dea7
Update Python environment installation command
