+++ .github/workflows/tests.yml
@@ -32,12 +32,12 @@ jobs:
     
     - name: Setup Python environment
       run: |
-        python3 -m pip install --upgrade build pip setuptools
+        python -m pip install --upgrade build pip setuptools
         pip install poetry
 
     - name: Install dependencies
       run: |
-        poetry install --with test
+        poetry install
 
     - name: Install OpenTerrace
       run: |
