+++ .github/workflows/tests.yml
@@ -2,41 +2,48 @@ name: Tests
 
 on:
   push:
-    branches:
+    branches: 
       - main
 
 jobs:
   test-pip:
     name: Run tests (with pip)
+
     runs-on: ${{ matrix.os }}
+    
     strategy:
       matrix:
         os: [ubuntu-latest]
         python-version: ['3.11']
+
     defaults:
       run:
         shell: bash
+
     steps:
-      - uses: actions/checkout@v2
+    - uses: actions/checkout@v2
 
-      - name: Set up Python ${{ matrix.python-version }} on ${{ matrix.os }}
-        uses: actions/setup-python@v2
-        with:
-          python-version: ${{ matrix.python-version }}
+    - run: git fetch --prune --unshallow --tags
 
-      - name: Setup Python environment
-        run: |
-          python -m pip install --upgrade pip setuptools
-          pip install poetry
+    - name: Set up Python ${{ matrix.python-version }} on ${{ matrix.os }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
 
-      - name: Install dependencies
-        run: poetry install
+    - name: Setup Python environment
+      run: |
+        python -m pip install --upgrade build pip setuptools
+        pip install poetry
 
-      - name: Install OpenTerrace
-        run: |
-          poetry build
-          pip install ./dist/openterrace-*.tar.gz
+    - name: Install dependencies
+      run: |
+        poetry install
 
-      - name: Run tests
-        run: |
-          poetry run pytest
+    - name: Install OpenTerrace
+      run: |
+        python3 -m build
+        pip3 freeze
+
+    - name: Run tests
+      run: |
+        python3 -m pytest
