### 2024-02-26 d78121d
Add pytest as a dependency for testing
