+++ pyproject.toml
@@ -27,6 +27,7 @@ numba = "*"
 scipy = "^1.9.3"
 matplotlib = "^3.6.1"
 tqdm = "^4.64.1"
+pytest = "^7.1.3"
 
 # Only used for documentation
 [tool.poetry.group.doc]
@@ -37,12 +38,6 @@ mkdocs-material = "*"
 mkdocstrings-python = "*"
 markdown-katex = "*"
 pytkdocs = "*"
-
-[tool.poetry.group.test]
-optional = true
-
-[tool.poetry.group.test.dependencies]
-pytest = "^7.1.3"
 
 [tool.poetry.group.github-actions]
 optional = true
