### 2024-02-26 9c020b2
Update build and test commands
