### 2024-02-26 9fb51e3
Update test command in workflow file
