+++ .github/workflows/tests.yml
@@ -46,4 +46,4 @@ jobs:
 
     - name: Run tests
       run: |
-        python -m pytest
+        pytest
