### 2024-02-26 e2373a8
Update test configuration and dependencies
