### 2024-02-27 05e98b8
Update GitHub Actions workflow for documentation deployment
