### 2024-02-27 588f913
Update test workflow to include multiple OS and Python versions
