+++ .github/workflows/tests.yml
@@ -13,8 +13,8 @@ jobs:
     
     strategy:
       matrix:
-        os: [ubuntu-latest]
-        python-version: ['3.11']
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.10,'3.11']
 
     defaults:
       run:
