### 2024-02-28 76c1b7e
Update schematic.svg and theory.md
