+++ docs/theory.md
@@ -1,8 +1,8 @@
 # Theory
 
 ## Schematic overview
-![Alt text](docs/_figures/schematic.svg)
-<img src="docs/_figures/schematic.svg">
+The fluid and bed phase are discretised in *N* and *M* nodes respectively. All bed material within the same fluid node is assumed to have the same temperature distribution.
+<img src="_figures/schematic.svg">
 
 ## Governing equations
 OpenTerrace solves partial differential equations for both phases.
