### 2024-02-28 e552a16
Update tutorials and save plots as SVG files
