+++ tutorials/tutorial5.py
@@ -54,4 +54,4 @@ plt.grid()
 plt.grid(which='major', color='#DDDDDD', linewidth=1)
 plt.grid(which='minor', color='#EEEEEE', linestyle=':', linewidth=0.8)
 plt.minorticks_on()
-plt.savefig('ot_plot_tutorial5.png')
+plt.savefig('ot_plot_tutorial5.svg')
