+++ docs/tutorials/tutorial4.md
@@ -1,0 +1,7 @@
+## Tutorial 4
+
+```python linenums="1"
+--8<-- "tutorials/tutorial4.py"
+```
+
+<img src="ot_plot_tutorial4.svg">
