+++ docs/tutorials/tutorial5.md
@@ -1,0 +1,7 @@
+## Tutorial 5
+
+```python linenums="1"
+--8<-- "tutorials/tutorial5.py"
+```
+
+<img src="ot_plot_tutorial5.svg">
