### 2024-02-28 01e2c9f
Update image links in documentation
