+++ docs/tutorials/tutorial1.md
@@ -1,7 +1,5 @@
-## Tutorial 1
-
 ```python linenums="1"
 --8<-- "tutorials/tutorial1.py"
 ```
 
-<img src="ot_plot_tutorial1.svg">
+[![](ot_plot_tutorial1.svg)](#)
