+++ .github/workflows/tests.yml
@@ -2,13 +2,14 @@ name: Tests
 
 on:
   push:
-    branches: 
+    branches:
       - main
+      - development
+  workflow_dispatch:
 
 jobs:
   test-pip:
     name: Run tests (with pip)
-
     runs-on: ${{ matrix.os }}
     
     strategy:
@@ -21,9 +22,11 @@ jobs:
         shell: bash
 
     steps:
-    - uses: actions/checkout@v2
+    - name: Checkout code
+      uses: actions/checkout@v3
 
-    - run: git fetch --prune --unshallow --tags
+    - name: Fetch all tags
+      run: git fetch --prune --unshallow --tags
 
     - name: Set up Python ${{ matrix.python-version }} on ${{ matrix.os }}
       uses: actions/setup-python@v2
