+++ .github/workflows/PyPi.yml
@@ -12,14 +12,17 @@ jobs:
       uses: actions/checkout@v3
       with:
         fetch-depth: 0
+
     - name: Set up Python
       uses: actions/setup-python@v3
       with:
         python-version: '3.11'
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install poetry        
+        pip install poetry
+             
     - name: Build and publish
       run: |
         poetry self add "poetry-dynamic-versioning[plugin]"
