+++ .github/workflows/docs.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches: 
       - main
+
 jobs:
   deploy:
     runs-on: ubuntu-latest
