### 2024-02-28 36e551c
Update test_compatibility.yml to use python directly instead of poetry run
