### 2024-02-28 5428089
Add tutorials to test_compatibility workflow
