### 2024-02-28 38699b3
Refactor simulation code and add main function
