### 2024-02-28 f4b361e
Update script entry points in pyproject.toml
