### 2024-02-28 e2de810
Add tutorials package to pyproject.toml
