+++ pyproject.toml
@@ -9,7 +9,8 @@ readme = "README.md"
 license = "MIT"
 
 packages = [
-    { include = "openterrace"}
+    { include = "openterrace"},
+    { include = "tutorials"}
 ]
 
 repository = "https://github.com/OpenTerrace/openterrace-python"
@@ -46,11 +47,11 @@ optional = true
 pytest-github-actions-annotate-failures = "^0.1.7"
 
 [tool.poetry.scripts]
-tut1 = "tutorials.tutorials1:main"
-tut2 = "tutorials.tutorials2:main"
-tut3 = "tutorials.tutorials3:main"
-tut4 = "tutorials.tutorials4:main"
-tut5 = "tutorials.tutorials5:main"
+tut1 = "tutorials1:main"
+tut2 = "tutorials2:main"
+tut3 = "tutorials3:main"
+tut4 = "tutorials4:main"
+tut5 = "tutorials5:main"
 
 [build-system]
 requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
