### 2024-02-28 3aba0e5
Remove test_compatibility.yml and update pyproject.toml
