+++ .github/workflows/test_compatibility.yml
@@ -1,55 +1,0 @@
-name: Test OS types and Python versions
-
-on:
-  push:
-    branches:
-      - development
-  workflow_dispatch:
-
-jobs:
-  test-pip:
-    name: Run tests (with pip)
-    runs-on: ${{ matrix.os }}
-    
-    strategy:
-      matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: ["3.11"]
-
-    defaults:
-      run:
-        shell: bash
-
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v3
-
-    - name: Fetch all tags
-      run: git fetch --prune --unshallow --tags
-
-    - name: Set up Python ${{ matrix.python-version }} on ${{ matrix.os }}
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Setup Python environment
-      run: |
-        python -m pip install --upgrade build pip setuptools
-        pip install poetry
-
-    - name: Install dependencies
-      run: |
-        poetry install
-
-    - name: Install OpenTerrace
-      run: |
-        python -m build
-        pip freeze
-
-    - name: Run tutorials
-      run: |
-        poetry run tut1
-        poetry run tut2
-        poetry run tut3
-        poetry run tut4
-        poetry run tut5
