### 2024-02-28 f691d96
Refactor function names to follow Python naming conventions
