### 2024-02-28 0c886ef
Update badges and Python version
