+++ pyproject.toml
@@ -21,7 +21,7 @@ vcs = "git"
 format = "v{base}"
 
 [tool.poetry.dependencies]
-python = "^3.11"
+python = "3.11"
 numpy = "^1.23.4"
 numba = "*"
 scipy = "^1.9.3"
