### 2024-02-28 81a8b01
Update Python version in pyproject.toml
