### 2024-02-28 e48fba1
Update GitHub Actions setup-python version
