+++ .github/workflows/upToDateDev.yml
@@ -12,7 +12,7 @@ jobs:
 
     steps:
     - name: Set Git config
-      uses: actions/checkout@v3
+      uses: actions/checkout@v4
 
     - name: Merge main back to dev
       run: |
