+++ mkdocs.yml
@@ -72,11 +72,11 @@ nav:
   - index.md
   - Installation guide: installation.md
   - Tutorials: 
-      - tutorials/tutorial1.md
-      - tutorials/tutorial2.md
-      - tutorials/tutorial3.md
-      - tutorials/tutorial4.md
-      - tutorials/tutorial5.md
+      - tutorials/tutorial_1.md
+      - tutorials/tutorial_2.md
+      - tutorials/tutorial_3.md
+      - tutorials/tutorial_4.md
+      - tutorials/tutorial_5.md
   - User guide:
       - user-guide/get-started.md
       - Theory: theory.md
