+++ docs/tutorials/tutorial4.md
@@ -1,5 +1,5 @@
 ```python linenums="1"
---8<-- "tutorials/tutorial4.py"
+--8<-- "tutorials/tutorial_4.py"
 ```
 
 [![](ot_plot_tutorial4.svg)](#)
