+++ docs/tutorials/tutorial3.md
@@ -1,5 +1,5 @@
 ```python linenums="1"
---8<-- "tutorials/tutorial3.py"
+--8<-- "tutorials/tutorial_3.py"
 ```
 
 [![](ot_plot_tutorial3.svg)](#)
