+++ docs/tutorials/tutorial1.md
@@ -1,5 +1,5 @@
 ```python linenums="1"
---8<-- "tutorials/tutorial1.py"
+--8<-- "tutorials/tutorial_1.py"
 ```
 
 [![](ot_plot_tutorial1.svg)](#)
