+++ docs/tutorials/tutorial5.md
@@ -1,4 +1,4 @@
 ```python linenums="1"
---8<-- "tutorials/tutorial5.py"
+--8<-- "tutorials/tutorial_5.py"
 ```
 [![](ot_plot_tutorial5.svg)](#)
