### 2024-02-28 2a59aad
Update docs.yml workflow to trigger on workflow_dispatch event
