+++ .github/workflows/docs.yml
@@ -2,8 +2,9 @@ name: Generate Docs
 
 on:
   push:
-    branches: 
+    branches:
       - main
+  workflow_dispatch:
 
 jobs:
   deploy:
