+++ docs/user-guide/get-started.md
@@ -1,4 +1,4 @@
 # Step-by-step guide
 
 1. Before getting started, please follow the [installation guide](../installation.md) to install OpenTerrace. If you plan to extend its capabilities it's recommended to install in [editable mode](../installation.md#install-in-editable-mode).
-2. Next, browse through the [tutorials](../tutorials.md) to get an idea of how to set up simulations.
+2. Next, browse through the [tutorials](../tutorials) to get an idea of how to set up simulations.
