+++ .github/workflows/test_numerical_schemes.yml
@@ -3,7 +3,7 @@ name: Test numerical schemes
 on:
   push:
     branches:
-      - development
+      - main
   workflow_dispatch:
 
 jobs:
