+++ docs/tutorials/tutorial_2.md
@@ -1,5 +1,5 @@
 ```python linenums="1"
---8<-- "tutorials/tutorial_2.py"
+--8<-- "tutorials/tutorial2.py"
 ```
 
 [![](ot_plot_tutorial2.svg)](#)
