### 2024-02-28 8307d14
Update version pattern in pyproject.toml
