+++ pyproject.toml
@@ -18,7 +18,7 @@ documentation = "https://openterrace.github.io/openterrace-python/"
 [tool.poetry-dynamic-versioning]
 enable = true
 vcs = "git"
-pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)(-?((?P<stage>[a-zA-Z]+)\\.?(?P<revision>\\d+)?))?"
+pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)"
 format-jinja = """
     {%- if distance == 0 -%}
         {{- base -}}
