### 2024-02-28 33f23c3
Update versioning format in pyproject.toml
