+++ pyproject.toml
@@ -18,14 +18,9 @@ documentation = "https://openterrace.github.io/openterrace-python/"
 [tool.poetry-dynamic-versioning]
 enable = true
 vcs = "git"
-pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)"
-format-jinja = """
-    {%- if distance == 0 -%}
-        {{- base -}}
-    {%- else -%}
-        {{- base }}.dev{{ distance }}+g{{commit}}
-    {%- endif -%}
-"""
+format = "v{base}"
+style = "pep440"
+latest-tag = true
 
 [tool.poetry.dependencies]
 python = "~3.11"
