### 2024-02-28 0bdfbbd
Update versioning format to use base value and semver style
