### 2024-02-28 8d66cb9
Update version pattern in pyproject.toml
