### 2024-02-28 08bd703
Update poetry-dynamic-versioning format in pyproject.toml
