### 2024-02-29 805d272
Update test-pip job to use poetry instead of pip
