### 2024-02-29 2c8de74
Fix Python version output in pytest workflow
