### 2024-02-29 1bc12e4
Update Python version setup in pytest workflow
