### 2024-02-29 3e07b10
Update pytest workflow to output Python version to GITHUB_OUTPUT
