### 2024-02-29 91c4f8a
Fix extracting Python version in pytest workflow
