+++ .github/workflows/pytest.yml
@@ -30,8 +30,8 @@ jobs:
     - name: Extract Python version and save as env variable
       id: python-version
       run: |
-      version=$(awk -F'=' '/^python/ {print $3}' "${{ github.workspace }}/${{ matrix.python-version-file }}")
-      echo "{version}=$version" >> $GITHUB_OUTPUT
+        version=$(awk -F'=' '/^python/ {print $3}' "${{ github.workspace }}/${{ matrix.python-version-file }}")
+        echo "{version}=$version" >> $GITHUB_OUTPUT
 
     - name: Set up Python ${{ env.version }} on ${{ matrix.os }}
       uses: actions/setup-python@v2
