### 2024-02-29 b2a0dde
Refactor Python version extraction in pytest.yml
