### 2024-02-29 9616178
Fix Python version extraction in pytest.yml
