### 2024-02-29 0dbb42e
Fix Python version extraction in pytest workflow
