### 2024-02-29 83efe5d
Refactor Python version extraction in pytest workflow
