+++ .github/workflows/pytest.yml
@@ -36,7 +36,7 @@ jobs:
         echo "version=$version" >> $GITHUB_ENV
 
     - name: Set up Python ${{ env.version }} on ${{ matrix.os }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ env.version }}
 
