### 2024-02-29 cb5697d
Update pytest workflow and badge in README
