### 2024-03-05 5a624a4
Update logo banner path in paper.md
