+++ tutorials/tutorial6.py
@@ -1,52 +1,58 @@
-"""
-This example sets up a water tank with PCM material
-as the bed material.
+""" 
+This example shows how a cylindrical pit storage looses
+energy to the surroundings over a period of 30 days.
 """
 
 import openterrace
 import matplotlib.pyplot as plt
+import numpy as np
+
+D = 2
+R = D/2
+H = 10
+T_inf = 273.15+10
 
 def main():
-    ot = openterrace.Simulate(t_end=100*60, dt=0.05)
+    ot = openterrace.Simulate(t_end=3600*24*30, dt=5)
 
-    fluid = ot.create_phase(n=100, type='fluid')
-    fluid.select_substance(substance='water')
-    fluid.select_domain_shape(domain='cylinder_1d', D=0.1, H=1)
-    fluid.select_porosity(phi=0.4)
-    fluid.select_schemes(diff='central_difference_1d', conv='upwind_1d')
-    fluid.select_initial_conditions(T=273.15+20)
-    fluid.select_massflow(mdot=0.01)
-    fluid.select_bc(bc_type='fixed_value',
-                    parameter='T',
-                    position=(slice(None, None, None), 0),
-                    value=273.15+80)
-    fluid.select_bc(bc_type='zero_gradient',
-                    parameter='T',
-                    position=(slice(None, None, None), -1))
-    fluid.select_output(times=[0, 30, 60, 90, 120, 150, 180, 210,
-                        240, 270, 300, 600, 900,
-                        1800, 3600, 5400, 6000])
+    fluid = ot.create_phase(n=101, type='fluid')
+    fluid.select_substance_on_the_fly(rho=1000, cp=4200, k=0.6)
+    fluid.select_domain_shape(domain='cylinder_1d', D=D, H=H)
+    fluid.select_schemes(diff='central_difference_1d')
+    fluid.select_initial_conditions(T=273.15+80)
+    fluid.select_bc(bc_type='zero_gradient', 
+                    parameter='T', 
+                    position=(slice(None, None, None), 0)
+                    )
+    fluid.select_bc(bc_type='zero_gradient', 
+                    parameter='T', 
+                    position=(slice(None, None, None), -1)
+                    )
+    fluid.select_output(times=range(0, 3600*24*60+3600*24, 3600*24*2))
 
-    bed = ot.create_phase(n=20, n_other=100, type='bed')
-    bed.select_substance('ATS58')
-    bed.select_domain_shape(domain='hollow_sphere_1d', Rinner=0.005, Router=0.025)
-    bed.select_schemes(diff='central_difference_1d')
-    bed.select_initial_conditions(T=273.15+20)
-    bed.select_bc(bc_type='zero_gradient',
-                parameter='T',
-                position=(slice(None, None, None), 0))
-    bed.select_bc(bc_type='zero_gradient',
-                parameter='T',
-                position=(slice(None, None, None), -1))
-    bed.select_output(times=range(0,600+300,300))
+    # Calculate surface area of each node
+    As = fluid.domain.dx*np.pi*D
 
-    ot.select_coupling(fluid_phase=0, bed_phase=1, h_exp='constant', h_value=200)
+    # Correct bottom node because its only half the height of the other nodes
+    As[0] = As[0]/2 + np.pi*(D/2)**2
+
+    # Correct top node because its only half the height of the other nodes
+    As[-1] = As[-1]/2 + np.pi*(D/2)**2
+
+    # Overall heat transfer coefficient
+    U = 10
+
+    # Thermal resistance
+    R = 1/(U*As)
+
[…]
