+++ docs/tutorials/tutorial_6.md
@@ -1,0 +1,5 @@
+```python linenums="1"
+--8<-- "tutorials/tutorial6.py"
+```
+
+[![](ot_plot_tutorial6.svg)](#)
