### 2024-03-06 cd96c2c
Add output_parameters argument to select_output method
