+++ openterrace/openterrace.py
@@ -268,17 +268,19 @@ class Simulate:
 
             self.sources.append({'R': R, 'T_inf': T_inf})
 
-        def select_output(self, times:list[float]=None):
+        def select_output(self, times:list[float]=None, output_parameters:list[str]=['T']):
             """Specify output times.
 
             Args:
                 times (float): List of times to output data
             """
+
+            self.output_parameters = output_parameters
             class Data(object):
                 pass
             self.data = Data()
             self.data.time = np.intersect1d(np.array(times), np.arange(self.outer.t_start, self.outer.t_end+self.outer.dt, self.outer.dt))
-            for parameter in ['T']:
+            for parameter in output_parameters:
                 setattr(self.data,parameter, np.full((len(self.data.time), self.n_other, self.n),np.nan))
                 self._flag_save_data = True
                 self._q = 0
@@ -292,7 +294,7 @@ class Simulate:
 
             if self._flag_save_data:
                 if t in self.data.time:
-                    for parameter in ['T']:
+                    for parameter in self.output_parameters:
                         getattr(self.data,parameter)[self._q] = getattr(self,parameter)
                     self._q = self._q+1
 
