+++ openterrace/tests/test_functions.py
@@ -2,187 +2,188 @@ import openterrace
 import numpy as np
 import matplotlib.pyplot as plt
 import datetime
+import pytest
+import sys
 
-class TestEnergyConservation:
-    def test_1(test):
-        n = 50
-        dt = 1e-2
-        t_end = 200
-        Lc = 0.025
-        T_init = 0
-        T_inf = 100
-        h = 200
-        cp = 4179
-        rho = 993
-        k = 0.627
+n = 11
 
-        ot = openterrace.Simulate(t_end=t_end, dt=dt)
-        bed2 = ot.create_phase(n=n, type='bed')
-        bed2.select_substance_on_the_fly(cp=cp, rho=rho, k=k)
-        #bed.select_domain_shape(domain='sphere_1d', R=Lc)
-        #bed.select_schemes(diff='central_difference_1d')
-        #bed.select_initial_conditions(T=T_init)
-        #bed.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), 0))
-        #bed.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), -1))
-        np.testing.assert_array_almost_equal(1,1, decimal=2) #Dummy check
+def test_1():
+    ot = openterrace.Simulate(t_end=100000, dt=1)
+    fluidbox = ot.create_phase(n=n, type='fluid')
+    fluidbox.select_substance_on_the_fly(cp=4200, rho=1000, k=0.6)
+    fluidbox.select_domain_shape(domain='block_1d', A=0.1, L=0.1)
+    fluidbox.select_schemes(diff='central_difference_1d')
+    fluidbox.select_initial_conditions(T=np.linspace(0,100,n))
+    fluidbox.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), 0))
+    fluidbox.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), -1))
+    fluidbox.select_output(times=range(0, 100000, 1000), output_parameters=['T','h'])
+    
+    ot.run_simulation()
 
-class TestDiffusion:
-    def test_1(self):
-        n = 50
-        dt = 1e-2
-        t_end = 200
-        Lc = 0.025
-        T_init = 0
-        T_inf = 100
-        h = 200
-        cp = 4179
-        rho = 993
-        k = 0.627
+    plt.plot(fluidbox.node_pos, fluidbox.data.h[:,0,:].T, label=fluidbox.data.time)
+    plt.legend(title='Simulation time (s)')
+    plt.xlabel(u'Position (m)')
+    plt.ylabel(u'Specific enthalpy (J/(kg))')
+    plt.grid()
+    plt.grid(which='major', color='#DDDDDD', linewidth=1)
+    plt.grid(which='minor', color='#EEEEEE', linestyle=':', linewidth=0.8)
+    plt.minorticks_on()
+    plt.savefig('ot_test_1.svg', bbox_inches='tight')
+    
+    np.testing.assert_array_almost_equal(1,1, decimal=2) #Dummy check
 
-        ot = openterrace.Simulate(t_end=t_end, dt=dt)
-        bed = ot.create_phase(n=n, type='bed')
-        bed.select_substance_on_the_fly(cp=cp, rho=rho, k=k)
-        bed.select_domain_shape(domain='sphere_1d', R=Lc)
-        bed.select_schemes(diff='central_difference_1d')
-        bed.select_initial_conditions(T=T_init)
[…]
