+++ .github/workflows/pytest.yml
@@ -51,4 +51,4 @@ jobs:
 
     - name: Run tests
       run: |
-        poetry run pytest
+        poetry run pytest -n auto
