### 2024-03-15 0560a09
Update coupling function and boundary conditions
