### 2024-03-21 #6411 @AdamRJensen
Hi again @baperry2 and @hewei8622

At JOSS we aim for reviewers to complete their review in 4-6 weeks, so I hope you're able to get started soon. I promise it's less daunting than it seems. Start by simply commenting on this PR with the text: `@editorialbot generate my checklist`
