### 2024-04-16 0122eb2
Refactor coupling function and boundary conditions in openterrace.py
