### 2024-04-17 OpenTerrace/openterrace-python#166 @baperry2
JOSS Review: Software Paper
This issue includes comments that should be addressed to resolve the "Software Paper" portion of the review checklist:

- [x] The summary describes the purpose of the software (studying thermal storage) but not the functionality provided by the software. A brief description of the uses of the software in studying thermal storage should be added.
- [x] The statement of need describes the intended uses and intended users of the software, but not the relation to other work. Is there other software used by researchers for similar purposes? Are there relevant publications of related scientific/engineering work that should be referenced to provide context?
- [x] State of the field: Not adequately addressed, can be resolved by answering the questions above in the paper.
- [x] Quality of writing: There is no problem with the quality of writing. However, the second figure should be referenced/described in the text and/or a caption should be provided so that the reader understands the significance.

The current references are good, but more may need to be added in resolving these points.
