### 2024-04-17 #6411 @baperry2
I opened https://github.com/OpenTerrace/openterrace-python/issues/166 that covers points related to the Paper portion of the checklist. I'll open additional issues and link here when I complete the remaining sections of the review.
