+++ docs/user-guide/verification-and-validation/verification.md
@@ -1,7 +1,10 @@
 # Verification studies
+The OpenTerrace code is tested against known analytical solution in limiting cases. The following presents such code verification results. These are run automatically as part of the [automated tests](https://github.com/OpenTerrace/openterrace-python/blob/main/.github/workflows/pytest.yml) using GitHub actions.
 
 ## Pure diffusion
 [![sphere diffusion](test_sphere_0.svg)](#)
 
 ## Pure advection
 To be added...
+
+![Python Version from PEP 621 TOML](https://img.shields.io/python/required-version-toml?tomlFilePath=https://raw.githubusercontent.com/OpenTerrace/openterrace-python/main/pyproject.toml)
