### 2024-04-22 731c435
Refactor central_difference_1d function in openterrace/diffusion_schemes/central_difference_1d.py
