### 2024-04-23 9de5d5d
Refactor test_5 function in test_functions.py
