### 2024-04-23 978d5fd
Update pyproject.toml with additional keywords and bug tracker URL
