### 2024-04-25 0e0c9b0
Update simulation duration in tutorial6.py
