+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'OpenTerrace: A pure Python framework for thermal energy storage packed bed simulations'
+title: 'OpenTerrace - A fast, flexible and extendable Python framework for thermal energy packed bed simulations'
 tags:
   - Python
   - Heat transfer
@@ -21,23 +21,27 @@ bibliography: paper.bib
 ![](https://raw.githubusercontent.com/OpenTerrace/openterrace-python/main/docs/_figures/logo-banner-paths-green.svg)
 
 # Summary
-Being able to store energy for extended periods of time is important for modern societies where increasing amounts of energy comes from renewable sources with time-varying production. Many storage technologies exist, each with their own set of disadvantages and advantages. Storing energy in the form of thermal energy is a promising solution because it is cheap and can easily be scaled up.
-
-`OpenTerrace` is a simulation framework to predict the transient temperature response of thermal energy storage systems. The thermal energy storage system contains a carrier fluid and an optional bed phase. Heat is transferred from the fluid to the bed phase during charging and from the bed phase to the fluid during discharging. `OpenTerrace` uses a finite volume formulation to discretise a set of coupled partial differential equations representing the fluid and bed phase. OpenTerrace comes with a wide range of predefined substances to be used as either fluid or bed material. Also, the set of primitive, predefined shapes that comes with `OpenTerrace` cover many applications. `OpenTerrace` is build to be:
+`OpenTerrace` is a simulation framework to predict the transient temperature response of thermal energy storage systems. The thermal energy storage system contains a carrier fluid and an optional bed phase. Heat is transferred from the fluid to the bed material during charging and from the bed phase to the fluid during discharging. `OpenTerrace` uses a finite volume formulation to discretise a set of coupled partial differential equations governing heat transfer in the fluid and bed phase. `OpenTerrace` comes with a wide range of predefined substances to be used as either fluid or bed material. Also, the set of primitive, predefined shapes that comes with `OpenTerrace` cover many applications. `OpenTerrace` is build to be:
 - Fast by making use of modern compilers and optimised tri-diagonal matrix solvers, OpenTerrace is built to be fast.
 - Flexible for easy integration in system models or optimisation loops.
 - Extendable by allowing new modules for new materials such as non-spherical rocks or exotic Phase Change Materials (PCM) to easily be plugged into the OpenTerrace framework.
 
-More information about how to get started and the latest functionaly can be in the [GitHub repository](https://github.com/OpenTerrace/openterrace-python) and the [Documentation](https://openterrace.github.io/openterrace-python/).
+More information about how to get started along with a user guide can be found in the [OpenTerrace documentation](https://openterrace.github.io/openterrace-python/). [Contribute](https://openterrace.github.io/openterrace-python/contributing/) to the [GitHub repository](https://github.com/OpenTerrace/openterrace-python) with new functionality by sending pull requests.
 
 # Statement of need
-`OpenTerrace` is an open-source simulation framework providing easy access to complex simulations. The core part of the computations are made by calling low-level languages using packages such as NumPy [@harris], SciPy [@virtanen] and  Numba [@lam] for speed while maintaining the flexibility and user-friendliness of Python. Plotting and animation capabilities based on Matplotlib [@hunter] are provided with the framework to allow easy access to visualisations.
+Being able to store energy for extended periods of time is important for modern societies where increasing amounts of energy comes from renewable sources with time-varying production. Many storage technologies exist, each with their own set of disadvantages and advantages. Storing energy in the form of thermal energy is a promising solution because it is cheap and can easily be scaled up.
 
-`OpenTerrace` solves for a fluid phase, which is descretised in `N` nodes. For each node, the bed material is discretised in `M` nodes. The framework allows arbritrary number of phases to be defined and coupled in different ways.
[…]
