+++ docs/contact.md
@@ -1,5 +1,8 @@
 # Contact
 
-Jakob Hærvig, Associate Professor, AAU Energy, Aalborg University
+Jakob Hærvig, Associate Professor (<jah@energy.aau.dk>)
 
-<jah@energy.aau.dk>
+Department of Energy (AAU Energy)<br />
+Pontoppidanstræde 111, 9220-Aalborg East<br />
+Aalborg University<br />
+Denmark
