+++ docs/students.md
@@ -1,8 +1,0 @@
-# Student projects
-Students are invited to help developing OpenTerrace.
-
-## Ideas for projects
-- Implement new discretisation schemes and benchmark them
-- Implement new correlations for e.g. convective heat transfer in porous media
-- Benchmark and optimise the code
-- Investigate new types of thermal storage systems by adding different fluids and or materials (also phase change materials)
