+++ openterrace/tests/test_functions.py
@@ -1,189 +1,0 @@
-import openterrace
-import numpy as np
-import matplotlib.pyplot as plt
-import datetime
-import pytest
-import sys
-
-
-
-def test_1():
-    ot = openterrace.Simulate(t_end=10000, dt=1)
-    fluidbox = ot.create_phase(n=101, type='fluid')
-    fluidbox.select_substance_on_the_fly(cp=4200, rho=1000, k=0.6)
-    fluidbox.select_domain_shape(domain='block_1d', A=0.1, L=0.1)
-    fluidbox.select_schemes(diff='central_difference_1d')
-    fluidbox.select_initial_conditions(T=np.linspace(0,100, 101))
-    fluidbox.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), 0))
-    fluidbox.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), -1))
-    fluidbox.select_output(times=range(0, 11000, 1000), output_parameters=['T','h'])
-    
-    ot.run_simulation()
-
-    plt.plot(fluidbox.node_pos, fluidbox.data.h[:,0,:].T, label=fluidbox.data.time)
-    plt.legend(title='Simulation time (s)')
-    plt.xlabel(u'Position (m)')
-    plt.ylabel(u'Specific enthalpy (J/(kg))')
-    plt.grid()
-    plt.grid(which='major', color='#DDDDDD', linewidth=1)
-    plt.grid(which='minor', color='#EEEEEE', linestyle=':', linewidth=0.8)
-    plt.minorticks_on()
-    plt.savefig('ot_test_1.svg', bbox_inches='tight')
-    
-    np.testing.assert_array_almost_equal(1,1, decimal=2) #Dummy check
-
-def test_2():
-    n = 50
-    dt = 1e-2
-    t_end = 200
-    Lc = 0.025
-    T_init = 0
-    T_inf = 100
-    h = 200
-    cp = 4179
-    rho = 993
-    k = 0.627
-
-    ot = openterrace.Simulate(t_end=t_end, dt=dt)
-    bed = ot.create_phase(n=n, type='bed')
-    bed.select_substance_on_the_fly(cp=cp, rho=rho, k=k)
-    bed.select_domain_shape(domain='sphere_1d', R=Lc)
-    bed.select_schemes(diff='central_difference_1d')
-    bed.select_initial_conditions(T=T_init)
-    bed.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), 0))
-    bed.select_bc(bc_type='zero_gradient', parameter='T', position=(slice(None, None, None), -1))
-
-    # Initialise array of thermal resistances
-    R = np.inf*np.ones_like(bed.T)
-
-    # Modify the thermal resistance at the surface of the sphere to account for convection
-    R[0][-1] = 1/(h*4*np.pi*Lc**2)
-
-    # Add the thermal resistance source term
-    bed.add_sourceterm_thermal_resistance(R=R, T_inf=T_inf)
-
-    ot.run_simulation()
-
-    Bi = h*Lc/k
-    Fo = k/(rho*cp)*t_end/Lc**2
-
-    r_r0_ana, theta_ana = openterrace.analytical_diffusion_sphere(Bi, Fo, n)
-    r_r0_num, theta_num = bed.node_pos/(bed.node_pos[-1]-bed.node_pos[0]), (bed.T[0,:]-T_inf)/(T_init-T_inf)
-
[…]
