+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'OpenTerrace - A fast, flexible and extendable Python framework for thermal energy storage packed bed simulations'
+title: 'OpenTerrace: A fast, flexible and extendable Python framework for thermal energy storage packed bed simulations'
 tags:
   - Python
   - Heat transfer
@@ -21,39 +21,34 @@ bibliography: paper.bib
 ![](https://raw.githubusercontent.com/OpenTerrace/openterrace-python/main/docs/_figures/logo-openterrace.svg)
 
 # Summary
-`OpenTerrace` is a simulation framework for prediction of transient temperature responses in thermal energy storage systems. The thermal energy storage system contains a carrier fluid and an optional bed phase. A set of coupled partial differential equations that govern heat transfer are solved using a finite volume formulation.
+`OpenTerrace` is a simulation framework for prediction of transient temperature responses in thermal energy storage systems. Being able to store energy for extended periods of time is important for modern societies where increasing amounts of energy stem from renewable sources with time-varying production. Many storage technologies exist, each with their own set of disadvantages and advantages. Storing energy in the form of thermal energy is a promising solution because it is cheap and can easily be scaled up. While the total energy content of thermal energy storage systems can be estimated easily, the transient response requires solving a set of partial differential equations in space and time. Using `OpenTerrace` the transient response of a wide range of thermal energy storage systems can be simulated. The storage system contains a storage tank filled with a carrier fluid and an optional bed phase. `OpenTerrace` comes with a wide range of predefined substances to be used as either fluid or bed material. Also, `OpenTerrace` ships with a set of primitive, predefined shapes that act as either storage tank and bed material. Besides that a set of predefined boundary conditions and source terms cover many different thermal energy storage applications.
 
-`OpenTerrace` comes with a wide range of predefined substances to be used as either fluid or bed material. Also, the set of primitive, predefined shapes that comes with `OpenTerrace` cover many thermal energy storage applications. `OpenTerrace` is build to be:
+`OpenTerrace` is build to be:
 - Fast by making use of modern compilers and optimised tri-diagonal matrix solvers, OpenTerrace is built to be fast.
 - Flexible for easy integration in system models or optimisation loops.
 - Extendable by allowing new modules for new materials such as non-spherical rocks or exotic Phase Change Materials (PCM) to easily be plugged into the OpenTerrace framework.
 
-More information about how to get started along with a user guide can be found in the [OpenTerrace documentation](https://openterrace.github.io/openterrace-python/). [Contribute](https://openterrace.github.io/openterrace-python/contributing/) to the [GitHub repository](https://github.com/OpenTerrace/openterrace-python) with new functionality by sending pull requests.
+More information about how to get started along with a user guide can be found in the [OpenTerrace documentation](https://openterrace.github.io/openterrace-python/). Users may send [pull requests](https://github.com/OpenTerrace/openterrace-python) to have their contributions with new functionality added to the official [OpenTerrace GitHub repository](https://github.com/OpenTerrace/openterrace-python). A set of tutorials is also provided with the framework to highlight its current functionality and to ease the learning curve for new users. Also, various unit tests are provided to verify different parts of the code in some well-defined benchmark cases.
 
 # Statement of need
-Being able to store energy for extended periods of time is important for modern societies where increasing amounts of energy comes from renewable sources with time-varying production. Many storage technologies exist, each with their own set of disadvantages and advantages. Storing energy in the form of thermal energy is a promising solution because it is cheap and can easily be scaled up.
[…]
