+++ paper/paper.bib
@@ -45,3 +45,37 @@
   location = {Austin, Texas},
   series = {LLVM '15}
 }
+
+@article{schumann,
+	author = {Schumann, T.E.W.},
+	title = {Heat transfer: {A} liquid flowing through a porous prism},
+	volume = {208},
+	doi = {10.1016/S0016-0032(29)91186-8},
+	number = {3},
+	journal = {Journal of the Franklin Institute},
+	year = {1929},
+	pages = {405--416}
+  }
+
+
+@article{littman,
+	title = {Gas-{Particle} {Heat} {Transfer} {Coefficients} in {Packed} {Beds} at {Low} {Reynolds} {Numbers}},
+	volume = {7},
+	doi = {10.1021/i160028a005},
+	number = {4},
+	journal = {Ind. Eng. Chem. Fundamen.},
+	author = {Littman, Howard and Barile, R G and Pulsifer, A H},
+	year = {1968}
+}
+
+
+@article{handley,
+	title = {The effect of thermal conductivity of the packing material on transient heat transfer in a fixed bed},
+	volume = {12},
+	doi = {10.1016/0017-9310(69)90038-6},
+	number = {5},
+	journal = {International Journal of Heat and Mass Transfer},
+	author = {Handley, D. and Heggs, P.J.},
+	year = {1969},
+	pages = {549--570}
+}
