### 2024-05-10 861733e
Update project description in pyproject.toml
