+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "openterrace"
-description = "OpenTerrace is a pure Python framework for thermal energy storage packed bed simulations"
+description = "OpenTerrace: A fast, flexible and extendable Python framework for thermal energy storage packed bed simulations"
 version = "0.0.0" # This will be overwritten by the dynamic versioning by Poetry
 authors = [
     "Jakob Hærvig"
