### 2024-05-10 OpenTerrace/openterrace-python#166 @jakobhaervig
All of the above is addressed in https://github.com/OpenTerrace/openterrace-python/commit/28f25c7f9466655a7c95dea5f8601da89e0c21f7. Can't find any open-source frameworks like OpenTerrace for thermal energy storage systems. Instead I added a small overview of relevant models from literature.
