### 2024-05-30 c28bd33
The conda environment creation command in the installation documentation has been updated to remove the specific Python version requirement. This allows for more flexibility in choosing the Python version for the environment.
