### 2024-05-30 651730b
Updated docs with new landing page for the user guide
