### 2024-05-31 253fc29
Update Python version in pytest workflow to test ^3.8
