+++ .github/workflows/pytest.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version-file: ['pyproject.toml']
+        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
 
     defaults:
       run:
@@ -29,23 +29,10 @@ jobs:
     - name: Fetch all tags
       run: git fetch --prune --unshallow --tags
 
-    # - name: Extract Python version and save as env variable
-    #   id: python-version
-    #   run: |
-    #     file_path="${{ github.workspace }}/${{ matrix.python-version-file }}"
-    #     version=$(awk -F'=' '/^python/ {print $2}' "$file_path" | tr -d '~' | xargs)
-    #     echo "Python version extracted: $version"
-    #     echo "version=$version" >> $GITHUB_ENV
-
-    # - name: Set up Python ${{ env.version }} on ${{ matrix.os }}
-    #   uses: actions/setup-python@v5
-    #   with:
-    #     python-version: ${{ env.version }}
-
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version-file: 'pyproject.toml'
+        python-version: ${{ matrix.python-version }}
 
     - name: Install dependencies
       run: |
