+++ pyproject.toml
@@ -37,7 +37,7 @@ format = "v{base}"
 latest-tag = true
 
 [tool.poetry.dependencies]
-python = "^3.11.8"
+python = "^3.8"
 numpy = "^1.26.4"
 numba = "^0.59.0"
 scipy = "^1.12.0"
