### 2024-05-31 25dad47
Update Python version in pytest workflow to test ^3.9
