### 2024-05-31 1fc371a
Python version in pytest workflow to include Python 3.9, 3.10, 3.11, and 3.12
