### 2024-06-02 #6411 @AdamRJensen
@jakobhaervig will you make a new release of the software and archive it on Zenodo (and subsequently write the doi here).
