### 2024-06-03 1c62a34
Update CITATION.cff with preferred citation information
