### 2024-06-03 #6411 @AdamRJensen
Awesome job @jakobhaervig getting in these proposed changes fast and being a pleasure to work with. Congratulations with your paper! 🥳

A particular a many thanks to you @baperry2 for your in-depth review. Your work is extremely appreciated and, in my view, very important for making true open science a reality!
