+++ README.md
@@ -4,7 +4,7 @@
 ![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/openterrace/openterrace-python/pytest.yml?label=tests)
 ![GitHub Actions Workflow Status](https://img.shields.io/github/actions/workflow/status/openterrace/openterrace-python/docs.yml?label=docs)
 [![GitHub Issues](https://img.shields.io/github/issues/openterrace/openterrace-python?label=github%20issues)](https://github.com/OpenTerrace/openterrace-python/issues)
-[[![status](https://joss.theoj.org/papers/cc89107c74cba69a948a98e5b0604336/status.svg)](https://doi.org/10.21105/joss.06411)
+[![status](https://joss.theoj.org/papers/cc89107c74cba69a948a98e5b0604336/status.svg)](https://doi.org/10.21105/joss.06411)
 
 ![Logo](https://raw.githubusercontent.com/OpenTerrace/openterrace-python/main/docs/_figures/logo-openterrace.svg)
 
