### 2024-06-04 #6411 @kyleniemeyer
Hi @jakobhaervig, I made some minor edits to the paper. Could you review and merge https://github.com/OpenTerrace/openterrace-python/pull/201 ?
