### 2024-06-04 #6411 @kyleniemeyer
@jakobhaervig it seems there is now an issue with the SVG figure, perhaps because I moved the caption into the figure definition. Could you convert it to a PDF (which would need to happen anyways in the PDF generation step) and replace that in the paper?
