### 2024-06-04 e20f455
Update logo file format in paper.md
